Update and rename Task.lua to task.lua
